+++ backend/controllers/user.controller.js
@@ -54,12 +54,18 @@ exports.updateUser = (req, res) => { /*Exports to the User router a updateUser()
                 likes: user.likes
             }
 
-            UserModel.updateOne({ _id: req.params.id }, { ...userObject, _id: req.params.id }) /*...then updates the former User object with the new information*/
-                .then(() => res.status(200).json({ message: 'La modification a été effectuée !' }))
-                .catch(error => {
-                    const errors = errorHandling(error)
-                    res.status(400).json({ errors })
+            const filename = user.picture.split('/profil/')[1] /*Otherwise it targets in the "images" folder any image associated with this user...*/
+
+            if (filename !== "random-user.png") {
+                fs.unlink(`../frontend/public/uploads/profil/${filename}`, () => {
+                    UserModel.updateOne({ _id: req.params.id }, { ...userObject, _id: req.params.id }) /*...then updates the former User object with the new information*/
+                        .then(() => res.status(200).json({ message: "La modification a été effectuée !" }))
+                        .catch(error => {
+                            const errors = errorHandling(error)
+                            res.status(400).json({ errors })
+                        })
                 })
+            }
         })
         .catch(error => res.status(400).json({ error }))
 }
@@ -77,11 +83,13 @@ exports.deleteUser = (req, res) => { /*Exports to the User router a deleteUser()
             }
 
             const filename = user.picture.split('/profil/')[1] /*Otherwise it targets in the "images" folder any image associated with this user...*/
-            fs.unlink(`../frontend/public/uploads/profil/${filename}`, () => { /*...then runs the FS unlink() function to delete this image from the folder...*/
-                UserModel.deleteOne({ _id: req.params.id }) /*...before deleting the User object itself*/
-                    .then(() => res.status(200).json({ message: `Le compte de ${user.pseudo} a été supprimé !` }))
-                    .catch(error => res.status(500).json({ error }))
-            })
+            if (filename && filename !== "random-user.png") {
+                fs.unlink(`../frontend/public/uploads/profil/${filename}`, () => { /*...then runs the FS unlink() function to delete this image from the folder...*/
+                    UserModel.deleteOne({ _id: req.params.id }) /*...before deleting the User object itself*/
+                        .then(() => res.status(200).json({ message: `Le compte de ${user.pseudo} a été supprimé !` }))
+                        .catch(error => res.status(500).json({ error }))
+                })
+            }
         })
         .catch(error => res.status(404).json({ error }))
 }
